- restore gibberish check from microcode_ctl.spec
Changed files:
    microcode-data-intel.spec -> 1.3

+++ microcode-data-intel.spec
@@ -11,19 +11,27 @@ Provides:	microcode-data
 ExclusiveArch:	i686 pentium2 pentium3 pentium4 x86_64 ia32e
 BuildRoot:	%{tmpdir}/%{name}-%{version}-root-%(id -u -n)
 
+# nothing to put there
+%define		_enable_debug_packages	0
+
 %description
 The microcode data file for Linux contains the latest microcode
 definitions for all Intel processors.
 
 %prep
 %setup -q -c
+mv microcode*.dat microcode.dat
 
 %build
+if ! grep -q 0x00000000 microcode.dat; then
+	echo >&2 microcode.dat contains giberrish
+	exit 1
+fi
+
 %install
 rm -rf $RPM_BUILD_ROOT
 install -d $RPM_BUILD_ROOT/lib/firmware
-
-install microcode*.dat $RPM_BUILD_ROOT/lib/firmware/microcode.dat
+cp -a microcode.dat $RPM_BUILD_ROOT/lib/firmware/microcode.dat
 
 %clean
 rm -rf $RPM_BUILD_ROOT
